requirements-dev.txt
# ════════════════════════════════════════════════════════════════
# DocuFlow - Requirements for Development & Local Testing
# Install with: pip install -r requirements-dev.txt
# ════════════════════════════════════════════════════════════════

# Core Data Processing & Science Stack
numpy>=1.26.4
pandas>=2.2
scikit-learn>=1.4
matplotlib==3.8.3
python-dateutil>=2.9

# Document & Image Processing
pillow==10.2.0
reportlab==4.1.0
opencv-python-headless==4.9.0.80
qrcode==7.4.2
PyMuPDF>=1.23.0

# Data Generation & Utilities
faker==28.4.1
tqdm==4.66.2
requests==2.31.0
editdistance==0.8.1
unidecode

# Validation & Standards
python-stdnum>=1.20

# Web Frameworks & APIs
flask>=3.0
gunicorn>=21.2

# Storage & Databases
minio==7.2.4
pymongo==4.6.1
python-dotenv==1.0.1

# Testing
pytest>=8.0
joblib>=1.3

# Development tools
ipython
black
flake8
